outposts: make listener IP configurable
